Moved image upload progress to separate plugin.
